requirements.txt
numpy
pandas
transformers
torch
trl
peft
datasets
huggingface-hub
pydantic
tqdm
Jinja2
pathlib
python-dotenv
